refactor: break up matrix
